--- src/utils/generate-label.ts
@@ -1,4 +1,0 @@
-export const createLabelerWithCount = (label: string, initCount = 0) => {
-  let count = initCount;
-  return () => `#${count++}-${label}`;
-};
