--- src/utils/gerenate-key.ts
@@ -1,1 +1,0 @@
-export const generateKey = () => Math.random().toString(36).substr(2, 11);
